Move tests to z_compound
