+++ contracts/test/vault/z_compound.js
@@ -1,4 +1,4 @@
-const { compoundVaultFixture } = require("../_fixture");
+const { defaultFixture, compoundVaultFixture } = require("../_fixture");
 const { expect } = require("chai");
 const { BigNumber, utils } = require("ethers");
 
@@ -670,5 +670,53 @@ describe("Vault with Compound strategy", function () {
         }
       }
     }
+  });
+});
+
+describe("Vault auto allocation", async () => {
+  if (isGanacheFork) {
+    this.timeout(0);
+  }
+
+  const mintDoesAllocate = async (amount) => {
+    const { anna, vault, usdc, governor } = await loadFixture(
+      compoundVaultFixture
+    );
+    await vault.connect(governor).setVaultBuffer(0);
+    await vault.allocate();
+    await usdc.connect(anna).mint(usdcUnits(amount));
+    await usdc.connect(anna).approve(vault.address, usdcUnits(amount));
+    await vault.connect(anna).mint(usdc.address, usdcUnits(amount));
+    return (await usdc.balanceOf(vault.address)).isZero();
+  };
+
+  const setThreshold = async (amount) => {
+    const { vault, governor } = await loadFixture(compoundVaultFixture);
+    await vault.connect(governor).setAutoAllocateThreshold(ousdUnits(amount));
+  };
+
+  it("Triggers auto allocation at the threshold", async () => {
+    await setThreshold("25000");
+    expect(await mintDoesAllocate("25000")).to.be.true;
+  });
+
+  it("Triggers auto allocation above the threshold", async () => {
+    await setThreshold("25000");
+    expect(await mintDoesAllocate("25001")).to.be.true;
+  });
+
+  it("Does not trigger auto allocation below the threshold", async () => {
+    await setThreshold("25000");
+    expect(await mintDoesAllocate("24999")).to.be.false;
+  });
+
+  it("Governer can change the threshold", async () => {
+    await setThreshold("25000");
+  });
+
+  it("Non-governer cannot change the threshold", async () => {
+    const { vault, anna } = await loadFixture(defaultFixture);
+    await expect(vault.connect(anna).setAutoAllocateThreshold(10000)).to.be
+      .reverted;
   });
 });
